requirements.txt
Django==2.1.4
django-crispy-forms==1.9.1
django-filter==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
mysqlclient==2.0.1
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
django-taggit